pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-document-assistant"
version = "0.1.0"
description = "A tool for processing documents using visual and text AI models via Ollama."
authors = [
  { name = "Muhammad Atif", email = "atifazad83@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "langchain",
  "langchain-community",
  "langchain-ollama",
  "faiss-cpu",
  "ollama",
  "fastapi",
  "pydantic",
  "uvicorn",
  "pypdf",
  "pdf2image",
  "pytesseract",
  "Pillow",
  "opencv-python",
  "numpy",
  "python-dotenv",
  "python-multipart",
  "pytest",
  "httpx",
  "requests",
  "black",
  "ruff"
]

[tool.uv]
# uv-specific settings can go here 